% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.compute.extinction.R
\name{ACDC.compute.extinction}
\alias{ACDC.compute.extinction}
\title{Create the piecewise-constant extinction rate from speciation rate}
\usage{
ACDC.compute.extinction(
  congruence.class,
  func.lambda = NULL,
  list.funcs.lambda = NULL
)
}
\arguments{
\item{congruence.class}{The model congruence class}

\item{func.lambda}{A speciation rate function}

\item{list.funcs.lambda}{A list of speciation rate functions}
}
\value{
An object with a list of speciation and extinction rate functions.
}
\description{
Create the piecewise-constant extinction rate from speciation rate
}
