% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.compute.speciation.R
\name{ACDC.compute.speciation}
\alias{ACDC.compute.speciation}
\title{Create the piecewise-constant speciation rate from extinction rate}
\usage{
ACDC.compute.speciation(congruence.class, func.mu = NULL, list.funcs.mu = NULL)
}
\arguments{
\item{congruence.class}{The model congruence class}

\item{func.mu}{A extinction rate function}

\item{list.funcs.mu}{A list of extinction rate functions}
}
\value{
An object with a list of speciation and extinction rate functions.
}
\description{
Create the piecewise-constant speciation rate from extinction rate
}
