% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.congruent.models.R
\name{ACDC.congruent.models}
\alias{ACDC.congruent.models}
\title{Create a set of congruent models}
\usage{
ACDC.congruent.models(model, mus = NULL, lambdas = NULL, keep_ref = TRUE)
}
\arguments{
\item{model}{The reference model. An object of class "ACDC"}

\item{mus}{A list of extinction-rate functions}

\item{lambdas}{A list of speciation-rate functions}

\item{keep_ref}{Whether or not to keep the reference model in the congruent set}
}
\value{
An object of class "ACDCset"
}
\description{
Create a set of congruent models
}
\examples{

lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

## A reference model
model <- congruence.class(lambda, mu, times = seq(0, 5, by = 0.005))

mu1 <- lapply(c(0.5, 1.5, 3.0), function(m) function(t) m)

model_set1 <- ACDC.congruent.models(model, mus = mu1)

model_set1

lambda0 = lambda(0.0) ## Speciation rates must all be equal at the present
bs <- c(-0.1, 0.0, 0.1)
lambda1 <- lapply(bs, function(b) function(t) lambda0 + b*t)

model_set2 <- ACDC.congruent.models(model, lambdas = lambda1)

model_set2
}
