% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.plot.congruence.class.R
\name{ACDC.plot.congruence.class.heatmap}
\alias{ACDC.plot.congruence.class.heatmap}
\title{Plots for congruent birth-death models.}
\usage{
ACDC.plot.congruence.class.heatmap(func_spec0, func_ext0, max.t, sample.grid)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{sample.grid}{Sampled congruent models to plot. Output of sample.congruence.class.}
}
\value{
A ggplot2 object.
}
\description{
Plots for congruent birth-death models.
}
\examples{
#TODO
}
