% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.plot.congruence.class.spaghetti.R
\name{ACDC.plot.congruence.class.spaghetti}
\alias{ACDC.plot.congruence.class.spaghetti}
\title{Plots for congruent birth-death models.}
\usage{
ACDC.plot.congruence.class.spaghetti(
  func_spec0,
  func_ext0,
  max.t,
  sample.grid,
  max.curves = 500
)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{sample.grid}{Sampled congruent models to plot. Output of sample.congruence.class.}

\item{max.curves}{Maximum number of curves to plot.}
}
\value{
No value returned, plot is generated.
}
\description{
Compared to make.congruence.class.plot, this plots individual curves of rates.
This allows vizualizing specific patterns, at the cost of either large files or smaller numbers of plotted curves.
}
\examples{
#TODO
}
