% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.plot.congruent.models.R
\name{ACDC.plot.congruent.models}
\alias{ACDC.plot.congruent.models}
\title{Plots the rate functions including the pulled rates.}
\usage{
ACDC.plot.congruent.models(obj)
}
\arguments{
\item{obj}{A list of rate functions produced by .}
}
\value{
void.
}
\description{
Plots the rate functions including the pulled rates.
}
\examples{
#TODO
}
