% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.read.R
\name{ACDC.read.RevBayes}
\alias{ACDC.read.RevBayes}
\title{Title}
\usage{
ACDC.read.RevBayes(path, times = seq(from = 0, to = 73, length.out = 1000))
}
\arguments{
\item{path}{path to log}

\item{times}{the time knots to be used in the piecewise-linear model}

\item{max_t}{tree height}
}
\value{
a list of piecewise-linear birth-death models, each being a sample in the posterior
}
\description{
Title
}
