% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.rates.R
\name{bootstrap.rates}
\alias{bootstrap.rates}
\title{Samples curves like the given curve by bootstrapping.}
\usage{
bootstrap.rates(num.epochs, func_rate, max.t, keep.rate0 = TRUE,
  reverse.time = FALSE, invert = FALSE, replace = TRUE,
  block.size = 1)
}
\arguments{
\item{num.epochs}{The discretization fineness}

\item{func_rate}{Function for getting rates at time t.}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{keep.rate0}{If TRUE, rate at present is fixed, otherwise chosen as func_rate at a random time.}

\item{reverse.time}{If TRUE, resampled curve is "backwards" from real curve, such that increases become decreases and vice-versa. Incompatible with keep.rate0=TRUE}

\item{invert}{If TRUE, returns a value proportional to 1/rates, allowing increases to become decreases while preserving rate at present.}

\item{replace}{If FALSE, only permutations of the curve are considered, if TRUE curves will be more variable.}

\item{block.size}{If replace=TRUE, larger values allow resampling to preserve larger-scale features of the rate trajectory.}
}
\value{
Rate at all changepoints
}
\description{
Samples curves like the given curve by bootstrapping.
}
