% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.pulled.speciation}
\alias{compute.pulled.speciation}
\title{Create the piecewise-constant pulled speciation rate}
\usage{
\method{compute}{pulled.speciation}(v_spec0, v_ext0, times)
}
\arguments{
\item{v_spec0}{The speciation rate at all change-points}

\item{v_ext0}{The extinction rate at all change-points}

\item{times}{The change-points}
}
\value{
Pulled speciation rate at all change-points
}
\description{
Create the piecewise-constant pulled speciation rate
}
\keyword{internal}
