% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.speciation.R
\name{compute.speciation}
\alias{compute.speciation}
\title{Create the piecewise-constant speciation rate}
\usage{
compute.speciation(lambda0, v_p_div, v_ext1, delta_t)
}
\arguments{
\item{lambda0}{The rate at present}

\item{v_p_div}{The pulled diversification rate at all changepioints}

\item{v_ext1}{The extinction rate at all changepoints}

\item{delta_t}{The width of each grid cell}
}
\value{
Speciation rate at all changepoints
}
\description{
Create the piecewise-constant speciation rate
}
\keyword{internal}
