% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruence.class.R
\name{congruence.class}
\alias{congruence.class}
\title{Computes the congruent class, i.e., the pulled rates.}
\usage{
congruence.class(func_spec0, func_ext0, max.t, num.intervals = 1000)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{num.intervals}{The number of discretization.}
}
\value{
A list of rate functions representing this congruence class.
}
\description{
Computes the congruent class, i.e., the pulled rates.
}
\examples{
#TODO
}
