% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.compute.speciation.R
\name{congruent.speciation}
\alias{congruent.speciation}
\title{Create the congruent model from extinction rate}
\usage{
congruent.speciation(model, func.mu)
}
\arguments{
\item{func.mu}{A extinction rate function}

\item{congruence.class}{The model congruence class}

\item{list.funcs.mu}{A list of extinction rate functions}
}
\value{
An object with a list of speciation and extinction rate functions.
}
\description{
Create the congruent model from extinction rate
}
