% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRABS.R
\docType{package}
\name{CRABS-package}
\alias{CRABS}
\alias{CRABS-package}
\title{CRABS: Congruent Rate Analyses in Birth-death Scenarios}
\description{
Features tools for exploring congruent phylogenetic birth-death models. It can construct the pulled speciation- and net-diversification rates from a reference model. Given alternative speciation- or extinction rates, it can construct new models that are congruent with the reference model. Functionality is included to sample new rate functions, and to visualize the distribution of one congruence class. See also Louca & Pennell (2020) \doi{10.1038/s41586-020-2176-1}.
}
\section{References}{


\itemize{
 \item Louca, S., & Pennell, M. W. (2020). Extant timetrees are consistent with a myriad of diversification histories. Nature, 580(7804), 502-505.


}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/afmagee/CRABS}
}

}
\author{
\strong{Maintainer}: Bjørn Tore Kopperud \email{kopperud@protonmail.com}

Authors:
\itemize{
  \item Sebastian Höhna
  \item Andrew F. Magee
}

}
