% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.congruent.rates.R
\name{FC}
\alias{FC}
\title{Fold change}
\usage{
FC(rates)
}
\arguments{
\item{rates}{Speciation/extinction rates at beginning/end of each piecewise linear interval.}
}
\value{
FC
}
\description{
max(rate(0),rate(t_max)) / min(max(rate(0),rate(t_max)))
}
\keyword{internal}
