% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.congruent.rates.R
\name{fnmean}
\alias{fnmean}
\title{Mean value of piecewise linear function}
\usage{
fnmean(rates)
}
\arguments{
\item{rates}{Speciation/extinction rates at beginning/end of each piecewise linear interval.}
}
\value{
mean
}
\description{
Mean value of piecewise linear function
}
\keyword{internal}
