% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.hsmrf.global.scale}
\alias{get.hsmrf.global.scale}
\title{Global scale for HSMRF using linear interpolation of pre-computed values}
\usage{
get.hsmrf.global.scale(v)
}
\arguments{
\item{v}{The number of pieces in the approximation}
}
\value{
Global scale
}
\description{
Global scale for HSMRF using linear interpolation of pre-computed values
}
\keyword{internal}
