% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruent.models.R
\name{joint.congruent.models}
\alias{joint.congruent.models}
\title{Create a set of congruent models}
\usage{
joint.congruent.models(model, mus, lambdas, keep_ref = TRUE)
}
\arguments{
\item{model}{The reference model. An object of class "CRABS"}

\item{mus}{A list of extinction-rate functions}

\item{lambdas}{A list of speciation-rate functions}

\item{keep_ref}{Whether or not to keep the reference model in the congruent set}
}
\value{
An object of class "CRABSset"
}
\description{
Create a set of congruent models
}
\examples{

# This function should not have to be used externally.
# It is called in the CRABS function `sample.congruence.class` when `rate.type=="joint"`.
}
