% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.congruent.rates.R
\name{MASV}
\alias{MASV}
\title{Mean absolute sequential variation}
\usage{
MASV(rates)
}
\arguments{
\item{rates}{Speciation/extinction rates at beginning/end of each piecewise linear interval.}
}
\value{
MASV
}
\description{
Measures net change.
}
\keyword{internal}
