% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDC.R
\name{plot.ACDC}
\alias{plot.ACDC}
\title{Plots the rate functions including the pulled rates.}
\usage{
\method{plot}{ACDC}(x, ...)
}
\arguments{
\item{x}{An object of class "ACDC"}

\item{...}{other parameters}
}
\description{
Plots the rate functions including the pulled rates.
}
\examples{
lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

model <- create.model(lambda, mu, times = seq(0, 5, by = 0.005))

plot(model)
}
