% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCsets.R
\name{plot.ACDCsets}
\alias{plot.ACDCsets}
\title{Plots the rate functions}
\usage{
\method{plot}{ACDCsets}(x, ...)
}
\arguments{
\item{x}{A list of objects of class ACDCsets}

\item{...}{other parameters}
}
\description{
Plots the rate functions
}
\examples{
lambda <- function(x) exp(0.3*x) - 0.5*x + 1
mu <- function(x) exp(0.3*x) - 0.2*x + 0.2
times <- seq(0, 5, by = 0.005)

model <- create.model(lambda, mu, times = times)

mus1 <- list(function(t) 0.2 + exp(0.1*t), 
            function(t) 0.2 + sin(0.35*t) + 0.1*t)
                        
mus2 <- list(function(t) 1.0, 
             function(t) 0.5 + 0.2*t)
                        
models <- list()
class(models) <- c("list", "ACDCsets") 
models[["set1"]] <- congruent.models(model, mus = mus1)

models[["set2"]] <- congruent.models(model, mus = mus2)

plot(models)

}
