% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CRABS.R
\name{plot.CRABS}
\alias{plot.CRABS}
\title{Plots the rate functions including the pulled rates.}
\usage{
\method{plot}{CRABS}(x, ...)
}
\arguments{
\item{x}{An object of class "CRABS"}

\item{...}{other parameters}
}
\value{
a patchwork object
}
\description{
Plots the rate functions including the pulled rates.
}
\examples{

data(primates_ebd)
lambda <- approxfun(primates_ebd$time, primates_ebd$lambda)
mu <- approxfun(primates_ebd$time, primates_ebd$mu)
times <- seq(0, max(primates_ebd$time), length.out = 500)

model <- create.model(lambda, mu, times = times)

plot(model)
}
