% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates}
\alias{primates}
\title{Primates phylogenetic tree}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(primates)
}
\description{
The example tree taken from the RevBayes tutorial website
}
\keyword{datasets}
