% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates_ebd_tess}
\alias{primates_ebd_tess}
\title{TESS Primates birth-death model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\usage{
data(primates_ebd_tess)
}
\description{
The results of a bayesian episodic birth-death model in the R-package TESS, fitted on the primates tree. One hundred episodes. Each estimate is the posterior median. The time unit is millions of years before the present.
}
\keyword{datasets}
