% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates_ebd_treepar}
\alias{primates_ebd_treepar}
\title{TreePar Primates birth-death model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\usage{
data(primates_ebd_treepar)
}
\description{
The results of a birth-death model in the R-package TreePar, fitted on the primates tree. The estimated model has two epochs, that are maximum-likelihood estimates. The time unit is millions of years before the present.
}
\keyword{datasets}
