% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDC.R
\name{print.ACDC}
\alias{print.ACDC}
\title{Print method for ACDC object}
\usage{
\method{print}{ACDC}(x, ...)
}
\arguments{
\item{x}{and object of class ACDC}

\item{...}{other arguments}
}
\description{
Print method for ACDC object
}
\examples{
data(primates_ebd)
lambda <- approxfun(primates_ebd$time, primates_ebd$lambda)
mu <- approxfun(primates_ebd$time, primates_ebd$mu)
times <- seq(0, max(primates_ebd$time), length.out = 500)

model <- create.model(lambda, mu, times = times)

print(model)
}
