% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCset.R
\name{print.ACDCset}
\alias{print.ACDCset}
\title{Print method for ACDCset object}
\usage{
\method{print}{ACDCset}(x, ...)
}
\arguments{
\item{x}{an object of class ACDCset}

\item{...}{other arguments}
}
\description{
Print method for ACDCset object
}
