% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCsets.R
\name{print.ACDCsets}
\alias{print.ACDCsets}
\title{Print method for ACDCsets object}
\usage{
\method{print}{ACDCsets}(x, ...)
}
\arguments{
\item{x}{and object of class ACDCsets}

\item{...}{other parameters}
}
\description{
Print method for ACDCsets object
}
