% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CRABS.R
\name{print.CRABS}
\alias{print.CRABS}
\title{Print method for CRABS object}
\usage{
\method{print}{CRABS}(x, ...)
}
\arguments{
\item{x}{and object of class CRABS}

\item{...}{other arguments}
}
\value{
nothing
}
\description{
Print method for CRABS object
}
\examples{
data(primates_ebd)
lambda <- approxfun(primates_ebd$time, primates_ebd$lambda)
mu <- approxfun(primates_ebd$time, primates_ebd$mu)
times <- seq(0, max(primates_ebd$time), length.out = 500)

model <- create.model(lambda, mu, times = times)

print(model)
}
