% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{One draw from a dirichlet distribution with concentration parameter alpha.}
\usage{
rdirichlet(dim, alpha)
}
\arguments{
\item{dim}{The dimension}

\item{alpha}{Concentration parameter}
}
\value{
Vector
}
\description{
One draw from a dirichlet distribution with concentration parameter alpha.
}
\keyword{internal}
