% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.congruence.class.R
\name{sample.congruence.class}
\alias{sample.congruence.class}
\title{Stochastic exploration of congruent models.}
\usage{
sample.congruence.class(
  model,
  num.samples,
  rate.type = "both",
  sample.speciation.rates = NULL,
  sample.extinction.rates = NULL
)
}
\arguments{
\item{model}{the reference model, an object of class "ACDC"}

\item{num.samples}{The pulled diversification rate function (measured in time before present).}

\item{rate.type}{either "extinction", "speciation", or "both"}

\item{sample.speciation.rates}{a function that when called returns a vector representing the speciation rate trajectory}

\item{sample.extinction.rates}{a function that when called returns a vector representing the extinction rate trajectory}
}
\value{
A named list with congruent rates.
}
\description{
Stochastic exploration of congruent models.
}
\examples{
#TODO
}
