% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.congruent.rates.R
\name{span}
\alias{span}
\title{Span of rates}
\usage{
span(rates)
}
\arguments{
\item{rates}{Speciation/extinction rates at beginning/end of each piecewise linear interval.}
}
\value{
span
}
\description{
max - min
}
\keyword{internal}
