% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.congruent.rates.R
\name{summarize.congruent.rates}
\alias{summarize.congruent.rates}
\title{Computes summary statistics for sampled congruent rates.}
\usage{
summarize.congruent.rates(func_spec0, func_ext0, max.t, sample.grid)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{sample.grid}{Sampled congruent models to plot. Output of sample.congruence.class.}
}
\value{
List of summaries for all congruent rates and observed rates.
}
\description{
Computes summary statistics for sampled congruent rates.
}
