% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.congruence.class.posterior.R
\name{summarize.posterior}
\alias{summarize.posterior}
\title{Title}
\usage{
summarize.posterior(
  posterior,
  threshold = 0.01,
  rate_name = "lambda",
  return_data = FALSE,
  rm_singleton = FALSE,
  relative_deltas = FALSE,
  ...
)
}
\arguments{
\item{threshold}{a threshold for when \eqn{\Delta\lambdai} should be interpreted as decreasing, flat, or increasing}

\item{rate_name}{either "lambda" or "mu"}

\item{return_data}{instead of plots, return the plotting dataframes}

\item{rm_singleton}{whether or not to remove singeltons. Pass starting at present, going towards ancient}

\item{relative_deltas}{whether to divide \eqn{\Delta\lambdai} by the local lambda value}
}
\value{

}
\description{
Title
}
