% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.trends.R
\name{summarize.trends}
\alias{summarize.trends}
\title{Title}
\usage{
summarize.trends(
  model_set,
  threshold = 0.005,
  rate_name = "lambda",
  return_data = FALSE,
  rm_singleton = FALSE,
  relative_deltas = FALSE,
  group_names = NULL
)
}
\arguments{
\item{model_set}{an object of type "ACDCset"}

\item{threshold}{a threshold for when \eqn{\Delta\lambdai} should be interpreted as decreasing, flat, or increasing}

\item{rate_name}{either "lambda" or "mu"}

\item{return_data}{instead of plots, return the plotting dataframes}

\item{rm_singleton}{whether or not to remove singeltons. Pass starting at present, going towards ancient}

\item{relative_deltas}{whether to divide \eqn{\Delta\lambdai} by the local lambda value}

\item{group_names}{a vector of prefixes, if you want to group the models in a facet. For example 'c("reference", "model")'}
}
\value{
a patchwork object
}
\description{
Title
}
\examples{

lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

reference <- create.model(lambda, mu, times = seq(0, 5, by = 0.005))

mu1 <- list(function(t) exp(0.2*t) - 0.3*t + 0.4,
function(t) exp(-0.8*t) - 0.1*t + 0.8,
function(t) exp(-1.5*t) + 0.2*t + 1.2,
function(t) 1.2 + 0.5*t,
function(t) 1.5 - 0.28*t)


model_set <- congruent.models(reference, mus = mu1)

p <- summarize.trends(model_set, 0.01)
}
