% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{acdc.loglikelihood}
\alias{acdc.loglikelihood}
\title{Compute likelihood}
\usage{
acdc.loglikelihood(model, phy, rho)
}
\arguments{
\item{model}{an object of class "ACDC"}

\item{phy}{an object of class "phylo"}

\item{rho}{the taxon sampling fraction}
}
\value{
the log-likelihood of the tree given the model
}
\description{
Compute likelihood
}
\examples{
lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

model <- create.model(lambda, mu, times = seq(0, 3, by = 0.005))

set.seed(123)
phy <- rcoal(25)

acdc.loglikelihood(phy, model)
}
