% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.read.RevBayes.R
\name{ACDC.read.RevBayes}
\alias{ACDC.read.RevBayes}
\title{Title}
\usage{
ACDC.read.RevBayes(
  path,
  max_t = 100,
  summary_type = "none",
  extinction_prefix = "extinction_rate.",
  speciation_prefix = "speciation_rate."
)
}
\arguments{
\item{path}{path to log}

\item{max_t}{tree height}

\item{summary_type}{either "none" for all the posterior samples, or "mean" or "median" for the posterior mean/median}

\item{extinction_prefix}{the prefix string for the extinction rate column names. Must be unique}

\item{speciation_prefix}{the prefix string for the speciation rate column names. Must be unique}
}
\value{
a set of ACDC models, each being a sample in the posterior
}
\description{
Title
}
\examples{

\dontrun{
#model <- ACDC.read.RevBayes("path/to/HSMRFBDP_primates.log", max_t = 73, )
}
}
