% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.compute.extinction.R
\name{compute.extinction}
\alias{compute.extinction}
\title{Create the piecewise-constant extinction rate}
\usage{
\method{compute}{extinction}(v_p_div, v_spec1, delta_t)
}
\arguments{
\item{v_p_div}{The pulled diversification rate at all change-points}

\item{delta_t}{The width of each grid cell}

\item{lambda0}{The rate at present}

\item{v_ext1}{The extinction rate at all change-points}
}
\value{
Extinction rate at all change-points
}
\description{
Create the piecewise-constant extinction rate
}
\keyword{internal}
