% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.pulled.diversification}
\alias{compute.pulled.diversification}
\title{Create the piecewise-constant pulled diversification rate}
\usage{
compute.pulled.diversification(v_spec0, v_ext0, delta_t)
}
\arguments{
\item{v_spec0}{The speciation rate at all changepioints}

\item{v_ext0}{The extinction rate at all changepoints}

\item{delta_t}{The width of each grid cell}
}
\value{
Pulled diversification rate at all changepoints
}
\description{
Create the piecewise-constant pulled diversification rate
}
\keyword{internal}
