% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruence.class.R
\name{congruence.class}
\alias{congruence.class}
\title{Computes the congruent class, i.e., the pulled rates.}
\usage{
congruence.class(
  func_spec0,
  func_ext0,
  times = seq(from = 0, to = 5, by = 0.005),
  func_p_spec = NULL,
  func_p_div = NULL
)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{num.intervals}{The number of discretization.}
}
\value{
A list of rate functions representing this congruence class.
}
\description{
Computes the congruent class, i.e., the pulled rates.
}
\examples{
lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

model <- congruence.class(lambda, mu, times = seq(0, 5, by = 0.005))

model
}
