% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.speciation.R
\name{congruent.speciation}
\alias{congruent.speciation}
\title{Create the congruent model from extinction rate}
\usage{
congruent.speciation(model, func.mu)
}
\arguments{
\item{model}{The reference model}

\item{func.mu}{An extinction rate function}
}
\value{
An object with a list of speciation and extinction rate functions.
}
\description{
Create the congruent model from extinction rate
}
