% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCset.R
\name{plot.ACDCset}
\alias{plot.ACDCset}
\title{Plots the rate functions}
\usage{
\method{plot}{ACDCset}(x, ...)
}
\arguments{
\item{x}{A list of congruent birth-death x}

\item{...}{other parameters}
}
\value{
nothing
}
\description{
Plots the rate functions
}
\examples{
lambda <- function(x) exp(0.3*x) - 0.5*x + 1
mu <- function(x) exp(0.3*x) - 0.2*x + 0.2
times <- seq(0, 5, by = 0.005)

model <- create.model(lambda, mu, times = times)

mus <- list(function(t) 0.2 + exp(0.1*t), 
           function(t) 0.2 + sin(0.35*t) + 0.1*t,
                       function(t) 1.0, 
                       function(t) 0.5 + 0.2*t)
models <- congruent.models(model, mus = mus)

plot(models)
}
