% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates_ebd_log}
\alias{primates_ebd_log}
\title{Primates birth-death model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 251 rows and 604 columns.
}
\usage{
data(primates_ebd_log)
}
\description{
See \code{?primates_ebd}, but including posterior samples instead of a summary.
}
\keyword{datasets}
