% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruence.class.R
\name{print.ACDC}
\alias{print.ACDC}
\title{Print method for ACDC object}
\usage{
\method{print}{ACDC}(x, ...)
}
\arguments{
\item{x}{and object of class ACDC}

\item{...}{other arguments}
}
\description{
Print method for ACDC object
}
\examples{
lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

model <- create.model(lambda, mu, times = seq(0, 5, by = 0.005))

print(model)
}
