% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruence.class.R
\name{print.ACDC}
\alias{print.ACDC}
\title{Print method for ACDC object}
\usage{
\method{print}{ACDC}(x, ...)
}
\arguments{
\item{x}{and object of class ACDC}

\item{...}{other arguments}
}
\description{
Print method for ACDC object
}
