% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruence.class.R
\name{print.ACDC}
\alias{print.ACDC}
\title{Print method for ACDC object}
\usage{
\method{print}{ACDC}(model, ...)
}
\arguments{
\item{model}{and object of class ACDC}

\item{...}{other arguments}
}
\value{

}
\description{
Print method for ACDC object
}
