% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.congruent.models.R
\name{print.ACDCset}
\alias{print.ACDCset}
\title{Print method for ACDCset object}
\usage{
\method{print}{ACDCset}(models, ...)
}
\arguments{
\item{models}{an object of class ACDCset}

\item{...}{other arguments}
}
\value{

}
\description{
Print method for ACDCset object
}
