% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCset.R
\name{print.ACDCset}
\alias{print.ACDCset}
\title{Print method for ACDCset object}
\usage{
\method{print}{ACDCset}(x, ...)
}
\arguments{
\item{x}{an object of class ACDCset}

\item{...}{other arguments}
}
\description{
Print method for ACDCset object
}
\examples{
lambda <- function(x) exp(0.3*x) - 0.5*x + 1
mu <- function(x) exp(0.3*x) - 0.2*x + 0.2
times <- seq(0, 5, by = 0.005)

model <- create.model(lambda, mu, times = times)

mus <- list(function(t) 0.2 + exp(0.1*t), 
           function(t) 0.2 + sin(0.35*t) + 0.1*t,
                       function(t) 1.0, 
                       function(t) 0.5 + 0.2*t)
models <- ACDC.congruent.models(model, mus = mus)

print(models)
}
