% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCsets.R
\name{print.ACDCsets}
\alias{print.ACDCsets}
\title{Print method for ACDCsets object}
\usage{
\method{print}{ACDCsets}(x, ...)
}
\arguments{
\item{x}{and object of class ACDCsets}

\item{...}{other parameters}
}
\description{
Print method for ACDCsets object
}
\examples{
lambda <- function(t) exp(0.3*t) - 0.5*t + 1
mu <- function(t) exp(0.3*t) - 0.2*t + 0.2

## A reference model
model <- create.model(lambda, mu, times = seq(0, 5, by = 0.005))

mu1 <- lapply(c(0.5, 1.5, 3.0), function(m) function(t) m)

model_set <- congruent.models(model, mus = mu1)
print(model_set)
}
