% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ACDCsets.R
\name{print.ACDCsets}
\alias{print.ACDCsets}
\title{Print method for ACDCsets object}
\usage{
\method{print}{ACDCsets}(model_sets, ...)
}
\arguments{
\item{...}{other arguments}

\item{model}{and object of class ACDCsets}
}
\value{

}
\description{
Print method for ACDCsets object
}
