% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.congruence.class.posterior.R
\name{sample.congruence.class.posterior}
\alias{sample.congruence.class.posterior}
\title{Title}
\usage{
sample.congruence.class.posterior(
  posterior,
  num.samples,
  rate.type = "both",
  sample.speciation.rates = NULL,
  sample.extinction.rates = NULL
)
}
\arguments{
\item{posterior}{a list of ACDC model objects}

\item{num.samples}{The pulled diversification rate function (measured in time before present).}

\item{rate.type}{either "extinction", "speciation", or "both"}

\item{sample.speciation.rates}{a function that when called returns a vector representing the speciation rate trajectory}

\item{sample.extinction.rates}{a function that when called returns a vector representing the extinction rate trajectory}
}
\value{
A named list with congruent rates.
}
\description{
Title
}
\examples{
data(primates_ebd_log)

posterior <- read.RevBayes(primates_ebd_log, max_t = 65, n_samples = 20)

extinction_rate_samples <- function(){
res <- sample.basic.models(
    num.epochs = 100,
    rate0.median = 0.1,
    model = "MRF",
    max.rate = 1.0)
  return(res)
}

samples <- sample.congruence.class.posterior(posterior, 
                                             num.samples = 20,
                                             rate.type = "extinction",
                                             sample.extinction.rates = extinction_rate_samples)

print(samples)
}
