% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.congruence.class.R
\name{sample.congruence.class}
\alias{sample.congruence.class}
\title{Stochastic exploration of congruent models.}
\usage{
sample.congruence.class(
  func_spec0 = NULL,
  func_ext0 = NULL,
  func_p_div = NULL,
  max.t,
  num.epochs,
  num.samples,
  rate.type = "both",
  sample.speciation.rates = NULL,
  sample.extinction.rates = NULL
)
}
\arguments{
\item{func_spec0}{The speciation rate function (measured in time before present).}

\item{func_ext0}{The extinction rate function (measured in time before present).}

\item{func_p_div}{The pulled diversification rate function (measured in time before present).}

\item{max.t}{The maximum time (before present) to consider rates.}

\item{num.epochs}{DESCRIPTION NEEDED}

\item{num.samples}{DESCRIPTION NEEDED}

\item{rate.type}{DESCRIPTION NEEDED}

\item{sample.speciation.rates}{DESCRIPTION NEEDED}

\item{sample.extinction.rates}{DESCRIPTION NEEDED}
}
\value{
A named list with congruent rates.
}
\description{
Stochastic exploration of congruent models.
}
\examples{
#TODO
}
