% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.rates.R
\name{sample.rates}
\alias{sample.rates}
\title{Sample custom functions through time.}
\usage{
sample.rates(num.epochs, lambda0 = NULL, rsample = NULL,
  rsample0 = NULL, autocorrelated = FALSE)
}
\arguments{
\item{num.epochs}{The number of rates to draw}

\item{lambda0}{The rate at present}

\item{v_p_div}{The pulled diversification rate at all changepioints}

\item{v_ext1}{The extinction rate at all changepoints}

\item{delta_t}{The width of each grid cell}
}
\value{
Extinction rate at all changepoints
}
\description{
Sample custom functions through time.
}
