% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.rates.R
\name{sample.rates}
\alias{sample.rates}
\title{Sample custom functions through time.}
\usage{
sample.rates(
  num.epochs,
  lambda0 = NULL,
  rsample = NULL,
  rsample0 = NULL,
  autocorrelated = FALSE
)
}
\arguments{
\item{num.epochs}{The number of rates to draw}

\item{lambda0}{The rate at present}

\item{rsample}{Function to sample next rate}

\item{rsample0}{Function to sample rate at present}

\item{autocorrelated}{Should rates be autocorrelated?}
}
\value{
Sampled rate vector
}
\description{
Sample custom functions through time.
}
