% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.rates.R
\name{sample.rates}
\alias{sample.rates}
\title{Sample custom functions through time.}
\usage{
sample.rates(
  num.epochs,
  lambda0 = NULL,
  rsample = NULL,
  rsample0 = NULL,
  autocorrelated = FALSE
)
}
\arguments{
\item{num.epochs}{The number of rates to draw}

\item{lambda0}{The rate at present}

\item{rsample}{Function to sample next rate}

\item{rsample0}{Function to sample rate at present}

\item{autocorrelated}{Should rates be autocorrelated?}
}
\value{
Sampled rate vector
}
\description{
Sample custom functions through time.
}
\examples{
data("primates_ebd")

l <- approxfun(primates_ebd[["time"]], primates_ebd[["lambda"]])
mu <- approxfun(primates_ebd[["time"]], primates_ebd[["mu"]])
model <- create.model(l, mu, primates_ebd[["time"]])

rsample <- function(n) runif(n, min = 0.0, max = 0.9)
mu_vals <- sample.rates(100, 0.5, rsample = rsample)
mu <- approxfun(model$times, mu_vals)

model_set <- congruent.models(model, mus = mu)

model_set
}
